+++ tests/all.ts
@@ -53,6 +53,7 @@ import { testit as issue5865 } from "./issue-5865.ts";
 import { testit as issue5866 } from "./issue-5866.ts";
 import { testit as issue5867 } from "./issue-5867.ts";
 import { testit as issue5868 } from "./issue-5868.ts";
+import { testit as issue5869 } from "./issue-5869.ts";
 import { testit as reloadDebounce } from "./reload-debounce.ts";
 import { testit as parseTipBrackets } from "./parse-tip-brackets.ts";
 import { testit as ghsaP2ph2rvmQ37m } from "./security-ghsa-p2ph-2rvm-q37m.ts";
@@ -102,6 +103,7 @@ const tests: [string, () => void | Promise<void>][] = [
   ["issue-5866", issue5866],
   ["issue-5867", issue5867],
   ["issue-5868", issue5868],
+  ["issue-5869", issue5869],
   ["reload-debounce", reloadDebounce],
   ["parse-tip-brackets", parseTipBrackets],
   ["security-ghsa-p2ph-2rvm-q37m", ghsaP2ph2rvmQ37m],